Update Split_Text.py
Once the database connection is established and data is inserted into the database, we fetch the data from the database executing the sql query. The result of the query is stored in the results variable which is then split into characters. Print i[1:30] is used to limit the number of characters that is being printed.